Enable assembly by default.
For init.c the flag is not set but the code
enables assembly if the flag is not set.
For the benchmarks, the code expects the flag
to be set and will not enable assembly by default.
Changing the build file allows the default for
benchmarks to be assembly enabled.

PiperOrigin-RevId: 273988816