Direct branch to source remainder handler for GEMM/IGEMM.
All ld64, ld128 and A53 microkernels move source remainder
to a branch forward (label 5) and branch directly to the
handler if there is less than main loop channels.

For 2 channels if there is a remainder, it is 1 channel so
additional check is not required.

For 4 channels if there is a remainder, and it is not 2 channels,
it is 1 channel so additional check is not required.

Standardize on label 4 for clamp code, 5 for remainder.
Should be a small performance improvement for 1 channel,
by branching directly to the remainder code,
and case where there is no remainder the branch is not
taken.

PiperOrigin-RevId: 294549181