Relax hswish FP16 microkernel test threshold
Was sporatic f16_hswish_test FAILED in F16_HSWISH__NEONFP16ARITH_X16.batch_div_16

The input range is -1 to 1, and likely inaccuracy near 0 is to blame.
Restricting the input range could be used instead.. typically 0.1 to 1
works better.  But this kernel tests negative values, so the input was left
unclamped.

PiperOrigin-RevId: 316064598