FP16 tests allow for results that are near zero.
When a float reference test produces a result that is not representable as a halffloat,
the halffloat result will be zero.  The tester tests for 1% error but will fail on
these near zero reference values.  So a minimum error of 0.001 is allowed.

This change increases the range of tests values to include near zero and in some cases
increases the accuracy requirement, but allows a minimum error.

PiperOrigin-RevId: 331656463