QS8 Neon GEMM C8 microkernel with 8 bit multiply and addvl sum the 8 values
C8 kernel using mull on 8 bit to 16 bit.  Then addlv to sum the 8 shorts to a single int.
After doing 4 of these, the result is packed into a single vector and accumulated with add.
The 4 int accumulators will represent 4 byte in the final output.
Comapred to partial sums, this uses 4 times fewer registers and no sum needed outside loop.

PiperOrigin-RevId: 354133661