QS8 Neon GEMM C8 microkernel with 8 bit multiply and vpadal to accumulate.
C8 partial sums kernel using mull on 8 bit to 16 bit a full 64 bits at a time.
Then padal to add pairs and lengthen to 32 bit, accumulating.
The 4 int accumulators will represent 1 byte in the final output, so there is
a vector for each element in the matrix.
The 4 ints are added together outside the loop.

PiperOrigin-RevId: 354631007