QS8 Neon GEMM C16 microkernel with two 8 bit multiplies and vpadal to accumulate.
C16 partial sums kernel using mull on 8 bit to 16 bit, then mlal on 8 bit to 16 bit
Then padal to add pairs and lengthen to 32 bit, accumulating.
The 4 int accumulators will represent 1 byte in the final output, so there is
a vector for each element in the matrix.
The 4 ints are added together outside the loop.

PiperOrigin-RevId: 358428739