Duplicate test methods in gemm-microkernel-test for JIT codegen, update IGEMM generator signature and test generation script.
Create new test methods in gemm-microkernel-tester for JIT codegen functions so that we can pass the generator min/max values for it to specialize. See src/f32-gemm/4x8-aarch32-neon-cortex-a53.cc for an example of how the jit_codegen_params is used.

The test generation script is also updated to not need to allocate buffer outside of calling Test (it is now done inside).

In order to differentiate between GEMM/IGEMM, we create a new type signature for IGEMM generators, which take an additional ks. Update all IGEMM JIT microkernels to have this new signature (not used yet).

PiperOrigin-RevId: 422865237