Integrate GEMM1, IGEMM, IGEMM1 into Convolution NHWC operator
Create helpers to try to allocate and generate JIT microkernels, fallback to fixed mircrokernels if that fails at any step.

Update f32-gemm-e2e to nullify either fixed or JIT microkernels to ensure we are testing the right thing. Consider when XNN_ENABLE_JIT is true, init.c will set the JIT generators, then inside convolution-nhwc, we will check if there is a generator available, and use it. This will be incorrect for non-JIT E2E benchmarks. If we guard the convolution-nhwc with XNN_ENABLE_JIT, then the JIT benchmarks will not run JIT microkernels unless we enable JIT.

PiperOrigin-RevId: 426504545