Create xnn_weights_cache that is a cache to hold repacked weights
xnn_weights_cache is a xnn_cache that holds xnn_weights_buffer. We have a separate buffer to hold weights as the allocation and free routines are different.
Add a simple init and release test, more functionality and tests will come later.

Rename some fields in xnn_cache to make it clearer if weight or code buffer is accessed.

PiperOrigin-RevId: 436760930