Add simple tests for convolution NHWC operator (GEMM microkernel) with weights cache
Add some test support for specifying to use weights cache. Right now we the tests just run with a weights cache, so we test that we can read and write packed weights into the cache. But since the weights are randomly generated, we won't exercise the cache hit case, that can be added in a follow-up.

More tests that exercise other gemm microkernels (igemm, etc) will be added later.

PiperOrigin-RevId: 437115090