Begin using weights cache in convolution-nhwc operator
Specify the operator's packed_weights field as a union. It is either a pointer (operator's weights_cache is NULL), or an offset into the weights cache (operator's weights_cache is not NULL).
Also add tests for convolution NHWC operator (GEMM microkernel) with weights cache. Add some test support for specifying to use weights cache. Right now we the tests just run with a weights cache, so we test that we can read and write packed weights into the cache. But since the weights are randomly generated, we won't exercise the cache hit case, that can be added in a follow-up.

PiperOrigin-RevId: 437351997