Introduce another kind of weights cache finalization that will leave extra space at the end of the memory
Previously, finalization means that any calls that can potentially change the weights cache (insert, reserve space) will be an error. Now we introduce a different kind of finalization that is less strict:

- We can reserve space on a finalized weights cache as long as it won't result in growing the memory
- we can get/insert into a weights cache as long as we have space in the cache for the incoming writes to the end of the cache's memory, and it is a cache hit (no insertion)

The weights cache will then track the maximum size of packed weights inserted into the cache, and reserve extra space when finalizing the memory. This allows a finalized weights cache to be reused, and as long as the packed weights are the same (i.e. same underlying model buffer, and thus always cache hits), we will not need to grow and move the weights memory.

The stricter finalization reduces memory usage, and is referred to as "hard". The looser one is referred to as "soft".

PiperOrigin-RevId: 443781097