Make xnn_workspace_t objects reference-counted
On the TFLite delegate side, both Delegate and Subgraph refers to workspace, but the lifetime of Delegate might be smaller than that of Subgraph, so we end up destroying the workspace (tied to delegate), before we delete the runtimes associated with the subgraphs, this causes segv in xnn_delete_runtime. Add a reference count to the workspace, which is incremented whenever a Delegate creates a workspace, and a runtime uses this. It is decremented whenever we call xnn_delete_workspace and xnn_delete_runtime, and actually deleted when reference count reaches 0.

PiperOrigin-RevId: 455498516