Fix ASAN failures in Max Pooling 2D subgraph tests
Output vector requires a buffer of XNN_EXTRA_BYTES.
Stride dimension also needs to be <= kernel dimension.
Restrict padding to be < kernel dimension so we don't get NaNs in output when comparing.
Change assertions of output to be restricted to output size without XNN_EXTRA_BYTES, so that we don't compare NaNs.

PiperOrigin-RevId: 456555408