Add an option to create weights cache specifying the capacity of the underlying buffer
By default, weights cache is 1MB. If the size of all weights used in a model exceeds 1MB, every operation that requires packing of weights will grow the weights cache by a small amount (size of the weights needed by the operator). This leads to unnecessary memory movements. We can increase the default size of the weights cache, but this penalizes models which don't use or have smaller weights. So, this new weights cache creation method takes a size, which is the initial capacity of the weights cache. The weights cache is guaranteed to not grow as long as the size of weights used in the weights cache is <= this initial size (though it can still grow once during soft finalization).

PiperOrigin-RevId: 463131786