Add in place operation optimization to memory planner and runtime
Some operations can be done in place (e.g. leaky relu, hard swish). In these cases, the output tensor can share the same memory as the input. We do this by marking the usage record of the output value with the input value's id. Then later, when setting the offset into the workspace, we check if this "reuse_value_id" is valid, if it is valid, we use the reuse_value_id's offset. The usage record of the input value needs to also be updated to expand to include consumers of the output value. We do not have an easy way of enumerating all consumers of a value, so we restrict this optimization to cases when value has a single consumer.

PiperOrigin-RevId: 471658308