Add in place operation optimization to memory planner and runtime
Some operations can be done in place (e.g. Leaky ReLU, Hard Swish). In these cases, the output tensor can share the same memory as the input.
We do this by marking the usage record of the output value with the input value's id and setting the usage record tensor size to 0.

By setting the size to 0, the planner does not check this tensor, and does not allocate any memory to it.

Later, after allocation of all tensors with size, we walk the entire list of usage records, and set alloc_offset of all tensors that are reusing space, by checking reuse_value_id.

The usage record of the input value needs to also be updated to expand to include consumers of the output value. We do not have an easy way of enumerating all consumers of a value, so we restrict this optimization to cases when value has a single consumer. In the future we can potentially add a value->last_consumer in order to expand this optimization.

Right now only unary elementwise operations, PReLU, Softmax, and Reshape are supported, binary elementwise operations can be added in the future.

PiperOrigin-RevId: 472778388