Hide asm symbols on __ELF__
Was there a reason why this wasn't done? I borrowed this file for IREE and noticed that my asm symbols were reported by nm as 'T' whereas all the C code around them was reported by nm as 't'. This fixes that. Also, it looks like the __MACH__ path just below does .private_extern, which is the equivalent?

PiperOrigin-RevId: 474714157