Fix QU8 DWCONV packing
The microkernels will read weights and subtract kernel_zero_point from it. This means that the "padding" weights due to kernel_size < primary_tile needs to be set to kernel_zero_point, otherwise the subtraction gives us negative values, instead of the expected 0.

We don't have any failing tests yet, because we don't yet test with kernel sizes < primary tile sizes. This will be added in a follow-up.

PiperOrigin-RevId: 481746723