Refactor conversion script and parse comment section
Also refactor parsing of prologue into separate function. The parsing is getting very nested and complicated, we split into a couple of steps:
- read input into list of strings
- split list into prologue and microkernel body
- parse prologue

We parse and store this information, but don't actually do anything with it right now, it will come in a follow-up patch.
The script now parses the comments in the assembly file, specifically the register usages. It will maintain a mapping of registers to their usages in a specific row in MR. This will allow us to emit guards for mr values when generating "upto" microkernels.

PiperOrigin-RevId: 486741890