Support converting assembly microkernels without support for post op
Introduce a new flag --post-op or --no-post-op (defaults to --post-op) for the converter. This will emitting any code to support post-operations. This is useful if the underlying microkernel cannot easily support post operations (e.g. the 4x8 A7 overwrites r5, and the post operation code loads from r5 assuming that it points to the params).

PiperOrigin-RevId: 487652024