Add dwconv microkernel tests and benchmarks for kernel size < primary tile
Also some fixes to step height calculation for benchmarks.

We probably want to rename kr to kernel_size, to make it clear it's not the tile size, but for this change I use kr to be kernel_size, and create a new primary_tile. This is because multipass uses kr for kernel size, we can fix it together in another change.

PiperOrigin-RevId: 492539665