Change packing to pack up to a rounding factor
Previously we would round channels up to subtile, but this would cause OOB reads for AVX, as the subtile is 8, and will lead to reading up to 7 elements OOB, but the max we can is 4 floats (determined by XNN_EXTRA_BYTES).
So, for such microarchitectures, we need to round channels up to just 4.
Packing is per normal, channel_tiles, followed by channel_subtiles.

This channel_round parameter is added to microkernel name.

PiperOrigin-RevId: 493407354