Generate F32 AVX and FMA3 DWCONV multipass microkernels
The weights are packed to channel_round == 4), so that for c == 30, we can do 2 loops of 16. We need to handle remainder (c % 8 != 0) in all passes (due to not being able to read > 4 elements out of bounds), e.g. c == 1.

Pad buffer with XNN_ALLOCATION_ALIGNMENT (so microkernels can write OOB to bufffer) and move XNN_ALLOCATION_ALIGNMENT into common.h to remove dependency on :allocator, and update callers to add the appropriate header.

PiperOrigin-RevId: 493657920