Add F32 to F16 DWCONV multipass weights packing functions and tests
Packing function is identical to existing F16/F32 multipass, except that we convert from fp32 to fp16 before writing to packed_weights.

Tests are also the same, and for easier visualization of the expected values (as float rather than fp16), we write a custom matcher to compare packed weights (fp16 as uint16_t) with expected floats.

PiperOrigin-RevId: 505833841