Fix dwconv unipass generator for even kernel sizes
For NEON, the current loop emits vmull when this conditions holds:

`K % 2 == 0 or K + 1 == KERNEL_TILE or not MLA:`

Currently:

KERNEL_SIZE == 4
K == 0 : vmull (K == 0)
K == 1 : vmlal (else)
K == 2 : vmull (K % 2 == 0)
K == 3 : vmull (K + 1 == KERNEL_SIZE)

This is wrong because we did not accumulate the result of K == 2.

The fix is to remove the `K + 1 == KERNEL_TILE` check, so that for even kernel sizes, the last K will be a mlal.

The vaddw is emitted correctly, as it is emitted when `K + 1 == KERNEL_TILE`.

Similar issue occurs for Wasm SIMD and SSE template.

Note, this does not change any generated unipass microkernels because they all have odd KERNEL_SIZE, so they always hit the `K % 2 == 0` when `K + 1 == KERNEL_TILE`.

PiperOrigin-RevId: 509932945