Fix incorrect type use for JIT FP16 GEMM in tests
The min/max values should be interpreted as float, and converted to fp16 when generating microkernels, however we were treating it as fp16. This did not lead to tests failing because this incorrect interpretation led to jit_gemm_params min and max value, so the generator would generate clamps. At runtime, we pass the correct min and max values, so the JIT code ends up doing the right thing.

PiperOrigin-RevId: 510257648