Fix runtime setup when one of operator objects is optimized away
A node can have multiple operators (e.g. even split). During runtime setup, if the operator object 0 is NULL, we skip setup for the entire node. This is wrong when we have a EvenSplit2 whose first output is optimized away and unused. We incorrectly skip setting up the copy operator for the second output.

We fix this by checking all operator objects in the node.

Also, even split is buggy as it indexes into operator object 0 in multiple places, which is a crash when we have an EvenSplit2 as described above.

Some naming clean up in the testers.

Also, no need to check that weights cache is finalized in the runtime setup code, because we check that in individual operator setup code.

Fixes #4318

PiperOrigin-RevId: 511812056