Add extra argument for DWCONV multipass packing
extra_bytes is used for packing scale values for QC8, this is usually channel_tile * sizeof(scale_type). However, this is incorrect in the channel_subtile loop, so we pass an additional value to indicate how many extra bytes to pad in the subtile loop of the last pass.

Also take this chance to consolidate signatures for unipass packing functions.

PiperOrigin-RevId: 513635209