Fix static values conversion to fp16 with multiple consumers
Nodes like convolution require that the weights and biases are fp32, and convert them to fp16 when packing. However, when these static values are also consumed by another node, e.g. a subtract, it marks the static value as fp16 compatible and converts them, then overwrites the value->data to the fp16 data. This leads to incorrect packing for convolution.

Thus, we keep the pointer to the original fp32 data around (in the fp32_data field), and the convolution node will check this field. If fp32_data is NULL, it will use data, otherwise, it uses fp32_data.

Note that this does not require any extra memory for static values as we do not allocate anything, we merely remember where the original fp32 data is (we do require 1 more sizeof(void*) field in xnn_value).

PiperOrigin-RevId: 514441437