Add a runtime flag to enable JIT
Previously, the compile time flag will include generators inside init.c. Lack of generators means JIT is disabled. But now, with a runtime flag, we use the presence of code cache to indicate if JIT is enabled or not. If it is NULL, JIT is disabled, otherwise, JIT is enabled.

The XNN_ENABLE_JIT compile option needs to be set in order for runtime flag to also work.

The main changes are in subgraph.h to make struct xnn_runtime take a pointer to code cache instead of holding the struct itself, so that it can be NULL.

PiperOrigin-RevId: 521893520