Invoke runtime in workspace tests to catch bugs in setup
xnn_setup_runtime sets value->data and operator data input/output pointers, which means that if workspace changes, the previously set up runtimes can have outdated pointers, even if the values have been updated (because the values are updated in initialize_workspace_values).

To check for such errors we call xnn_invoke_runtime, which will actually try to access the pointers set in operator data, and tests will fail if we did not update pointers properly

This actually catches 2 tests where we did not update the pointers correctly:

1. create shared worksapce
2. create runtime 1
3. setup runtime 1, size workspace for runtime 1
4. create runtime 2, bigger workspace needed, create new buffer, and update runtime1 values
5. setup runtime 2
6. invoke runtime 1
7. invoke runtime 2

Step 4 is incomplete because we only updated runtime1 values, but the pointers inside operator objects have already be set in step 3, so in step 6 we get asan errors as the workspace created in 1 has already been freed.

The 2 failing tests are disabled for now, a follow-up patch will fix this issue and re-enable the tests.

PiperOrigin-RevId: 525842210