Track workspace changes and update runtime if needed
We track if workspace has changed (only when it grows due to a creating a runtime which requires a larger workspace). If it has, we set a field in the workspace. During setup, we check this field, and update all runtimes that share this workspace (only those that have been set up before), by calling setup again, to re-initialize all the pointers.

It is wasteful to call setup again only to reinitialize the input/output pointers, we can probably add a small helper function that only does setting of pointers and call that.

PiperOrigin-RevId: 526633262