Fix value allocation type for fp16 rewrite
New values created for fp16 rewrite for external values should have their allocation type set to workspace, and not external (which is copied from the fp32 external value). Also, in runtime, we change the allocation type to dynamic for static fp32 values that we converted to fp16, but we did not check if value was static. An external value is fp16 compatible, and we will incorrectly set it to be dynamic allocation.

Added some new tests and checks to catch this, and also this fixes breakage.

PiperOrigin-RevId: 529147990