Fix wasm_assembler_test_web_simd_chrome-linux test failure
The two test-cases failing are those testing the behaviour under a large number of functions and a large number of local variables. They are failing due to 4K size-limit imposed by browsers on the bytecode being compiled in the main thread.

This decreases the definition of "large" from 300 to 129, which is large enough for the purpose of a test, as LEB128 encoding of unsigned 129 is two-byte long.

Also this fixes the "large number of local variables" test case, so that it now
relies on kLargeNubmerOfLocals, not kLargeNumberOfFunctions.

PiperOrigin-RevId: 537898914