Avoid using a temporary buffer for emitting WASM sections
WASM sections begin with a declaration of their size, which in turn is encoded with a variable-length format LEB128.
Previously, to simplify the matters, the section bodies were emitted to a buffer
first and only then copied to the result.

Now we precalculate the size and emit the sections directly.

PiperOrigin-RevId: 538831601