Deduplicate func types in WASM Assembler
WASM binary format allows to specify a function signature (functype) only once
in the Type section and then refer to it multiple times to define functions.
Previously, WASM assembler was just encoding a new functype for each function.
Functype deduplication allows to reduce the codegen size for modules with multiple
versions of the same function. In particular this increases
`kLargeNumberOfFunctions` from 129 to 235, yet the test still passes when run in
browser, in spite of the 4Kb limit imposed by browsers on the bytecode size
being compiled in the main thread.

This also adds a dedicated structure to model the function return type, clearly
limiting the number of returned values to one.

PiperOrigin-RevId: 539572209