Avoid dynamic allocations in ValTypesToInt
ValTypesToInt, that was used to pass the number of local variables to WASM assembler
used to be backed by std::vector, which is unjustified, as the container cannot contain more than 5 elements (by the number of types WASM defined).
Here we extract the generic part from LocalsArray and call it `ArrayPrefix` -- this container is backed by `std::array` and is now used to back ValTypesToInt.

PiperOrigin-RevId: 543432721