Avoid dynamic allocation when generating function bodies in WASM Assembler
We used to store the function body in a std::vector. Now we store it in the same buffer and move it to the right position upon finalisation. This avoids dynamic allocation and significantly cuts the size of the binary.

Also, this disables ManyFunctionsTestSuite, as it will be completely rewritten in the next CL.

PiperOrigin-RevId: 548703687