Provide uniform abstraction for assemblers to store into the code buffer
Due to the WASM binary format WASM assembler has to be able to write at
an arbitrary position of the code buffer. Previously this was implemented via
`Code` providing the necessary functionality itself, largely duplicating the
machinery provided by `BaseAssembler`. This extends the latter interface, which
allows to deduplicate code.

This is also an important preliminary step on the way to allow assemblers grow the
code buffer as needed.

PiperOrigin-RevId: 550525210