Add pointer types to WASM Assembler DSL
This adds `Ptr<Value>` wrapper over `Local`, that represents a pointer that
is aware of the type it points to, which simplifies pointer arithmetics.

In WebAssembly all pointers are represented
as `i32`. They don't "know" what they're pointing to, so for
`float* ptr` we could write `ptr += 2`, but for `Local ptr`, we have to write
`ptr = I32Add(ptr, I32Const(2 * sizeof(float)))`, which is not only cumbersome,
but has also proven to be the breeding ground for tricky bugs.

This also makes it necessary to allow the lambda fed into `AddFunc` to have arguments of both `Local` and `Ptr<V>` type.
This leads to the need to prepare the arguments as `std::tuple`, not `std::array`.

This layer of abstraction exists only at compile time and the compiler yields the
same code it did before.

PiperOrigin-RevId: 553160033