Unify declaration of arguments with function definition
We use restrict and array of min size 1 at the definition, but the declaration uses a pointer. This results in a lot of -Warray-parameter= warnings in the Bazel build. Unify the qualifiers used.

restrict keyword is only in C99, not in C++ and not supported on MSVC, so we guard it behind a XNN definition in common.h.

PiperOrigin-RevId: 565601259