Fix Bazel build for bench subdir
Several targets were including a nonexistent file in `srcs`, meaning they would not build properly. The file they (apparently) want to include is already in the deps set, so simple removal of these srcs fixes the build. (Perhaps this is a wart that appeared from a Google -> GitHub change migration?)