Fix reading of quantization parameters in GEMM microkernels.
The GEMM microkernel were currently reading past the end of the `quantization_params` data when `mr` is  smaller than the kernel's number of rows. This was not problematic in most cases since the quantization parameters are padded to avoid segmentation faults and/or uninitialized data, but did cause problems for batched GEMMs where the quantization parameters cannot be padded correctly.

This change uses the same "trick" as for the input and output pointers, i.e. explicitly mapping pointers to the data of the last rows to other valid rows.

PiperOrigin-RevId: 618848906