Add `copy_bias` helper functions and use it to refactor packing.c.
This also changes packing.c to be C++. This is the first of several refactors that I think can dramatically reduce the volume of code (and thus potential bugs), while at the same time making this much faster in a platform independent way.

I have a particular workload that speeds up over 2x end-to-end (with a pack function getting 8x faster) just by using memcpy where possible in this file. I'm not saying we can or should replace microkernels with code like this, but why not make reference code as fast as possible without compromising (or even improving) readability and maintainability?

PiperOrigin-RevId: 642324657