Use a better approximation for the `f32-vgelu` kernels.
The previous implementation used a 10/8-degree rational polynomial that minimized the absolute error, whereas this change uses a slightly higher-degree 12/10 rational polynomial that minimizes the relative error.

The new 12/10 kernels are 10-20% slower than the 10/8 kernels, but still significantly faster (~15x for `neon`, ~20x for `sse2`, ~55x for `fma3`, and ~85x for `avx512f`) than the naive version (scalar using `erff()`).

PiperOrigin-RevId: 650148922