Add a `ADD_SHARDED_TEST` macro to `CMakeLists.txt` to split up tests that are slow for `riscv64`, which runs on `qemu`.
This does not make the tests any faster, but allows them to be parallelized much better (fewer and shorter stragglers), roughly halving the time of the `cmake-linux-riscv-rvv` workflow.

This change also refactors `generate-vunary-test.py` so that `xnn_init_hardware_config()` is not repeatedly called in the main loop for `riscv` tests.

PiperOrigin-RevId: 652862760