Remove JIT support from XNNPACK
We believe this is unused, and not enabled by default. The motivation for removing it is that it's a lot of complex code that requires effort to maintain, which only makes sense if it is useful. There are currently broken tests that are going unfixed due to lack of maintenance/attention.

PiperOrigin-RevId: 662952914