Replace .yaml for unary and binary ops with headers that "call" C macros.
This allows replacing a bunch of per-microkernel code with macro calls:
- Declaring the microkernel headers (vunary.h, ...)
- Generated test/benchmark code only needs to know which microkernel header to include, rather than the .yaml spec

This works by making a header *without* include guards for each microkernel type, that calls a macro for each microkernel of that type. The includer defines the macro to do whatever it wants to do with each microkernel declaration (declare the function, instantiate a test/benchmark, etc.).

This doesn't remove the .yaml files yet, because I use them to generate the headers in this change. Once this change is merged, the .yamls can be removed (or I can remove them in this change just before submitting).

What I like about this change:
- Microkernels and associated metadata are enumerated in exactly one place (`{op}/microkernels.h`). Edit this one list, and you get everything else (microkernel declarations, tests, benchmarks), *without* regenerating any code.
- Similar to the above, this builds all the conditional compilation for ukernels in one place.
- Eliminates a huge amount of generated code. I think C preprocessor generated code is much better than python generated code: it can't get stale, and editing it is *much* easier than untangling the conditional code generation in the python templates. I was able to unwind a lot of tricky conditional logic in this change.

PiperOrigin-RevId: 667522243