Run benchmarks as tests
A lot of benchmarks are/were broken, we need to cover them with testing (not just building). This change enables the benchmarks to run as tests, with the `--benchmark_min_time=1x` argument, which only runs each benchmark once, to make test latency reasonable.

I've tagged many of them `manual` (and `notap`), because they are either way too slow, or crashing.

Fixed a bunch more benchmark bugs found when enabling these tests:
- rsum benchmarks either swapped rows/channels, or used the value of `channels` for both parameters, leading to huge benchmark latency and timeouts.
- Memory corruption in vcvt benchmark due to batch size being bytes, not elements and using the output and not input.
- deconvolution failed to allocate XNN_EXTRA_BYTES

PiperOrigin-RevId: 675800299