Use header tables for dwconv kernels
This doesn't touch benchmarks yet, because they are not generated currently, and have a lot of issues that need to be resolved first.

Some of the dwconv kernels were missing listing in the yaml files, hence not getting test coverage, including some that were used in production configs. This issue was detected automatically by this change, and this change fixes the problem.

PiperOrigin-RevId: 676602059