Add variant of FC which can take fp16 weights and biases and f32 inputs and outputs.
Some models have float16 weights to reduce model size. Currently we have to convert float16 weights to float32 within the XNNPack delegate and this memory is not freed until program termination. Also, if we are doing float 16 inference, we convert the f16 weights to f32 inside the delegate, and then back again in XNNPack.

So this will reduce memory usage for f32 inference since only one copy of f32 weights is active at any time. And for f16 inference, no conversion to f32 is ever done.

PiperOrigin-RevId: 677119973