Replace (some) end2end benchmarks with subgraph benchmarks
The existing "end2end" benchmarks are not actually end-to-end, because they skip the subgraph API, which contains nontrivial optimizations and work.

This change removes these existing end2end benchmarks, and adds subgraph benchmarks instead. Currently, only qs8, fp16, and fp32 mobilenet benchmarks are included. More datatypes and kinds of models are still TODO.

This also allows closing a const correctness loophole in the configs that existed specifically to enable these benchmarks to hack the configs.

There is also coverage of the individual ops at the microkernel level.

PiperOrigin-RevId: 677129403