Stop initializing test and benchmark outputs
This costs some time (important on emulators), and it's not useful IMO. If for some reason we are ever in a debugging exercise where this would be useful, it would be very easy to bring it back anyways, either temporarily or otherwise.

This was done with a find-replace: std::fill(.*output.*);\n -> (empty string), and then reverting some cases that were needed (e.g. initializing reference results for accumulators).

PiperOrigin-RevId: 684343387