Fix build identifier generation on Windows.
Windows builds fail on the generate_build_identifier genrule because the
command line is too long when expanding the source file paths to feed to the
Python generation script. To work around this, we write those files to a
single file that we.

This is what the new `xnnpack_source_list_file` rule does. It ONLY exposes the
list file as an output.

When the python script gets executed, the files it has access to are limited
to what is explicitly requested in the `srcs` attribute. We add the "list"
and the source paths to the sources to make them available to the Python
script and only expand the list file path to keep the command line short
enough to avoid having Windows break down.

I have not found a way to transitively pull the source files from the "list"
target and make them available from the Python script without adding them to
the `srcs` attribute (making them expanded in the "$(SRCS)" variable), thus
the repetition.

PiperOrigin-RevId: 684457906