Replace xnn_node_type_* for unary and binary elementwise ops
This eliminates the need for enumerating all unary and binary elementwise node types in a bunch of places.

This does result in a minor regression: when errors are reported, they are now less specific than they used to be. However, many of those more specific reports were wrong (as revealed by this change), e.g. reporting `rope` as `subtract`, and `space_to_depth_2d` as `clamp`.

PiperOrigin-RevId: 690526458