Optimize f16/bf16 abs, sign, negate reference kernels
Since we're removing bf16-vabs, I thought we might as well do this. It should make code size smaller too. I don't think we actually need to worry about the performance of this kernel, but it happens to generate code that performs equally well as the microkernel. Unsurprisingly, this is an easy case for the compiler.

PiperOrigin-RevId: 692098109