Improve confusing and less accurate computation of quantization parameters
The code currently computes the scale of the QD8 quantization parameters as 1.0f / (1.0f / scale), and then stores 1.0f / scale. This extra reciprocal throws away precision.

The fp16 version of this had a workaround for this problem, probably because this same problem with fp16 would be a really destructive loss in precision.

This change uses the same approach used for fp16 to avoid this double reciprocal. It also just makes things more consistent/less confusing.

PiperOrigin-RevId: 699088487