Use `benchmark::utils::WipeCache()` instead of relying on a set of ops with distinct buffers.
This clears up any issues with internal buffers, e.g. cached weights, which could be reused across operator runs.

Note that this only clears the L2 cache (if the size can be obtained), as clearing all caches with `benchmark::utils::WipeCache()` makes running the benchmarks (and not necessarily the benchmarked code) unreasonably slow.

PiperOrigin-RevId: 700693054